docs: review - 파일 위치 변경
- dto 이하 디렉토리에 req, res 디렉토리 추가
