+++ src/main/java/com/ioteam/order_management_platform/review/dto/req/CreateReviewRequestDto.java
@@ -1,4 +1,4 @@
-package com.ioteam.order_management_platform.review.dto;
+package com.ioteam.order_management_platform.review.dto.req;
 
 import java.util.UUID;
