+++ src/main/java/com/ioteam/order_management_platform/review/dto/res/ReviewResponseDto.java
@@ -1,10 +1,9 @@
-package com.ioteam.order_management_platform.review.dto;
+package com.ioteam.order_management_platform.review.dto.res;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 import com.ioteam.order_management_platform.review.entity.Review;
-import com.ioteam.order_management_platform.user.entity.User;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -13,7 +12,7 @@ import lombok.Getter;
 @Builder
 @AllArgsConstructor
 @Getter
-public class AdminReviewResponseDto {
+public class ReviewResponseDto {
 
 	private final UUID reviewId;
 	private final ReviewUserResponseDto user;
@@ -24,33 +23,24 @@ public class AdminReviewResponseDto {
 	private final UUID reviewOrderId;
 	private final Boolean isPublic;
 	private final LocalDateTime createdAt;
-	private final UUID createdBy;
-	private final LocalDateTime modifiedAt;
-	private final UUID modifiedBy;
-	private final LocalDateTime deletedAt;
-	private final UUID deletedBy;
 
-	public static AdminReviewResponseDto from(Review review, User user) {
-		return AdminReviewResponseDto
+	public static ReviewResponseDto from(Review review) {
+		return ReviewResponseDto
 			.builder()
 			.reviewId(review.getReviewId())
 			.user(ReviewUserResponseDto.from(review.getUser()))
+			.restaurant(ReviewRestaurantResponseDto.from(review.getRestaurant()))
 			.reviewScore(review.getReviewScore())
 			.reviewContent(review.getReviewContent())
 			.reviewImageUrl(review.getReviewImageUrl())
 			.isPublic(review.getIsPublic())
 			.createdAt(review.getCreatedAt())
-			.createdBy(review.getCreatedBy())
-			.modifiedAt(review.getModifiedAt())
-			.modifiedBy(review.getModifiedBy())
-			.deletedAt(review.getDeletedAt())
-			.deletedBy(review.getDeletedBy())
 			.build();
 	}
 
-	public static AdminReviewResponseDto from(Review review, ReviewUserResponseDto reviewUser,
+	public static ReviewResponseDto from(Review review, ReviewUserResponseDto reviewUser,
 		ReviewRestaurantResponseDto reviewRestaurant) {
-		return AdminReviewResponseDto
+		return ReviewResponseDto
 			.builder()
 			.reviewId(review.getReviewId())
 			.user(reviewUser)
@@ -60,11 +50,6 @@ public class AdminReviewResponseDto {
 			.reviewImageUrl(review.getReviewImageUrl())
 			.isPublic(review.getIsPublic())
 			.createdAt(review.getCreatedAt())
-			.createdBy(review.getCreatedBy())
-			.modifiedAt(review.getModifiedAt())
-			.modifiedBy(review.getModifiedBy())
-			.deletedAt(review.getDeletedAt())
-			.deletedBy(review.getDeletedBy())
 			.build();
 	}
 }
